requirements.txt
flask
functools
parse
pyserial
pyyaml
socket
systemd
vcgencmd
